requirements.txt
# Core Python packages
pygame
websockets
websocket-client
numpy

# Image processing and display
Pillow

# GUI framework
PyQt6
PySide6

# Raspberry Pi specific requirements (only install on Linux/Raspberry Pi)
gpiozero; platform_system == "Linux"
adafruit-circuitpython-ssd1306; platform_system == "Linux"
adafruit-blinka; platform_system == "Linux"
lgpio; platform_system == "Linux"

# Development and testing (optional)
pytest
